Add struct for the view key and moved the decrypt function
