+++ wasm/src/account/mod.rs
@@ -16,3 +16,6 @@
 
 pub mod private_key;
 pub use private_key::*;
+
+pub mod view_key;
+pub use view_key::*;
